Don't rely on unmanaged function pointers
Translating managed arguments to unmanaged arguments would be required for
this to actually work as intended, which would in turn require interpreting
the custom marshalling behaviors.

Instead, just invoke the managed delegate from the  method. Any
performance gain of the  instruction is probably unimportant when
invoking from managed code anyway. If it is important, the end-user can
always call  and cast to the appropriate unmanaged
function pointer type themselves.