Remove erroneous compile-time constant CallingConvention
Also note that UnmanagedCallConvs attribute was removed from Invoke during
refactoring. This was intentional. Getting a function pointer to Invoke is
not a supported scenario. Instead, unsafe code can use GetFunctionPointer and
cast to the appropriate type.

This further removes the need to interpret CallingConvention.Winapi. This
value is supported by the UnmanagedFunctionPointer attribute used on the
generated delegate.