Add NGD1002 diagnostic for CallingConvention
This diagnostic requires the value of the CallingConvention parameter in the
factory methods to be a literal enum value (variables, const fields, and
readonly fields are not supported). const fields and readonly fields could be
supported, but there's little to no practical benefit of doing so as the
field would still have to be initialized unconditionally in a way we could
parse.