Remove generic method handling for interface methods
Currently the interface methods do not declare any generic methods, yet there
remains explicit handling for these cases. A future version may support
generic methods for unmanaged function pointers (requiring  to be
enabled), but those methods can defer to the non-generic
 method directly. No other generic methods are foreseen
to be added to the interfaces at this time.